presentation-api/src/requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
boto3==1.24.22
botocore==1.27.22
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
exif==1.3.5
expiringdict==1.2.2
fastapi==0.78.0
ffmpeg-python==0.2.0
filetype==1.2.0
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
idna==3.3
iiif-prezi3==0.0.2.2
jmespath==1.0.1
mangum==0.15.0
pi==0.1.2
Pillow==9.2.0
plum-py==0.8.1
pydantic==1.9.1
pymongo==4.1.1
python-dateutil==2.8.2
python-magic==0.4.27
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.18.2
webencodings==0.5.1
